Update Sphinx dependency version
Updated the sphinx dependency version, and the basepython version
in order to fix the tox docs job. The job was no longer able to
run properly since the sphinx versions for python2 were outdated.
Now, we are using the openstack sphinx version requirements pattern,
and forcing the job running under python3, which fixes the sphinx
version issue.

Change-Id: Ib61a515c99d2e72f8d611612c639b6bbfca072d9